# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Финпульс — API учёта личных и семейных финансов
  version: 1.0.0
  description: |
    Контракт backend-системы учёта личных и семейных финансов.

    Архитектура: модульный монолит + выделенные сервисы **auth** и **reports**,
    за единой точкой входа (API Gateway). Все защищённые эндпоинты используют
    JWT (Bearer). Пользователь видит только свои данные и данные семей, в которых состоит.

    Ролевая модель внутри семьи: `OWNER` (админ), `MEMBER` (участник), `VIEWER` (просмотр).
  contact:
    name: Финпульс
servers:
  - url: http://localhost:8080/api/v1
    description: Локальный API Gateway
...
paths:
...
  /reports/summary:
    get:
      tags: [Reports]
      summary: Сводка за период (доход, расход, баланс, накопления)
      operationId: reportSummary
      parameters:
        - $ref: '#/components/parameters/FamilyIdQuery'
        - $ref: '#/components/parameters/PeriodFrom'
        - $ref: '#/components/parameters/PeriodTo'
        - $ref: '#/components/parameters/MemberIds'
      responses:
        '200':
          description: Сводка
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SummaryReport' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/by-category:
    get:
      tags: [Reports]
      summary: Структура по категориям
      operationId: reportByCategory
      parameters:
        - $ref: '#/components/parameters/FamilyIdQuery'
        - $ref: '#/components/parameters/PeriodFrom'
        - $ref: '#/components/parameters/PeriodTo'
        - $ref: '#/components/parameters/MemberIds'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/OperationType' }
      responses:
        '200':
          description: Доли по категориям
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CategorySlice' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/by-member:
    get:
      tags: [Reports]
      summary: Структура по участникам
      operationId: reportByMember
      parameters:
        - $ref: '#/components/parameters/FamilyIdQuery'
        - $ref: '#/components/parameters/PeriodFrom'
        - $ref: '#/components/parameters/PeriodTo'
        - name: type
          in: query
          schema: { $ref: '#/components/schemas/OperationType' }
      responses:
        '200':
          description: Суммы по участникам
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MemberSlice' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/cashflow:
    get:
      tags: [Reports]
      summary: Денежный поток за период с группировкой по дню/неделе/месяцу
      description: |
        Возвращает доходы и расходы, сгруппированные по интервалам внутри произвольного
        периода `from`..`to`. Для коротких периодов используйте мелкую гранулярность
        (`DAY`/`WEEK`), для года — `MONTH`.
      operationId: reportCashflow
      parameters:
        - $ref: '#/components/parameters/FamilyIdQuery'
        - $ref: '#/components/parameters/PeriodFrom'
        - $ref: '#/components/parameters/PeriodTo'
        - $ref: '#/components/parameters/MemberIds'
        - name: granularity
          in: query
          description: Шаг группировки. По умолчанию MONTH.
          schema:
            $ref: '#/components/schemas/Granularity'
      responses:
        '200':
          description: Поток по интервалам
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CashflowPoint' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /reports/export:
    get:
      tags: [Reports]
      summary: Выгрузка отчёта в файл
      operationId: exportReport
      parameters:
        - $ref: '#/components/parameters/FamilyIdQuery'
        - $ref: '#/components/parameters/PeriodFrom'
        - $ref: '#/components/parameters/PeriodTo'
        - $ref: '#/components/parameters/MemberIds'
        - name: format
          in: query
          required: true
          schema: { type: string, enum: [xlsx, pdf, csv] }
      responses:
        '200':
          description: Файл-документ отчёта
          content:
            application/octet-stream:
              schema: { type: string, format: binary }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    FamilyIdQuery:
      name: familyId
      in: query
      required: true
      schema: { type: string, format: uuid }
    PeriodFrom:
      name: from
      in: query
      required: true
      description: Начало периода (включительно). Период произвольный — месяц, квартал, год или любой диапазон дат.
      schema: { type: string, format: date, example: '2026-04-01' }
    PeriodTo:
      name: to
      in: query
      required: true
      description: Конец периода (включительно). Должен быть не раньше from.
      schema: { type: string, format: date, example: '2026-06-30' }
    MemberIds:
      name: memberIds
      in: query
      description: Список userId участников; пусто = вся семья
      schema:
        type: array
        items: { type: string, format: uuid }
      style: form
      explode: true
  responses:
...
    Forbidden:
      description: Нет доступа к ресурсу
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    OperationType:
      type: string
      enum: [INCOME, EXPENSE]
    Granularity:
      type: string
      enum: [DAY, WEEK, MONTH]
      default: MONTH
      description: Шаг группировки данных во времени для отчёта о денежном потоке
...
    SummaryReport:
      type: object
      properties:
        from: { type: string, format: date }
        to: { type: string, format: date }
        income: { type: number, format: double }
        expense: { type: number, format: double }
        balance: { type: number, format: double }
        savings: { type: number, format: double }
        savingsRate: { type: number, format: double, description: Доля сбережений 0..1 }
        incomeChangePct: { type: number, format: double, nullable: true }
    CategorySlice:
      type: object
      properties:
        categoryId: { type: string, format: uuid }
        name: { type: string }
        icon: { type: string }
        color: { type: string }
        amount: { type: number, format: double }
        percent: { type: number, format: double, description: Доля 0..100 }
    MemberSlice:
      type: object
      properties:
        userId: { type: string, format: uuid }
        name: { type: string }
        color: { type: string }
        amount: { type: number, format: double }
        percent: { type: number, format: double }
    CashflowPoint:
      type: object
      properties:
        month:
          type: string
          description: Ключ интервала — YYYY-MM для MONTH, YYYY-MM-DD (дата начала) для DAY/WEEK
          example: '2026-06'
        label: { type: string, description: Подпись интервала для оси графика, example: 'Июн' }
        income: { type: number, format: double }
        expense: { type: number, format: double }
...
    Error:
      type: object
      description: Единый формат ошибки (RFC-7807-подобный)
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        code: { type: string, example: OPERATION_NOT_FOUND }
        message: { type: string, example: Операция не найдена }
        path: { type: string }
      required: [status, code, message]
...
security:
  - bearerAuth: []